src/Azure.Cost.Importer/requirements.txt
azure-common==1.1.27
azure-core==1.20.1
azure-eventhub==5.6.1
azure-eventhub-checkpointstoreblob-aio==1.1.4
azure-graphrbac==0.61.1
azure-identity==1.7.1
azure-mgmt-authorization==0.61.0
azure-mgmt-containerregistry==8.2.0
azure-mgmt-core==1.3.0
azure-mgmt-keyvault==9.3.0
azure-mgmt-resource==13.0.0
azure-mgmt-storage==11.2.0
azure-monitor-opentelemetry-exporter==1.0.0b5
azure-storage-blob==12.10.0
pandas==1.3.4
requests==2.26.0
